requirements.txt
Django<6
djangorestframework
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
celery
redis
reportlab